fixed ui, better typing for AddExchange input
